Create ChatGPT-PE
Able to design, develop, debug, technology choices, and enhance feature capabilities using ChatGPT3.5 in 5-6 days. As long as we know what we need, understand the design and technology, how to read and understand the code, and can connect the dots, we can generate a basic infrastructure and framework in a matter of few days. Our prompting (Prompt Engineering Techniques) are very important.